Moving plotting functions to its own class and also adding manager as parameter to some TimeMachineUtils functions
