+++ utils/Functions.java
@@ -20,40 +20,4 @@ public class Functions {
         }
         return ret;
     }
-
-    public static Figure plot(double[][] x, double[][] y, String[] traceLabels, String xLabel, String yLabel) {
-        ScatterTrace[] traces = new ScatterTrace[x.length];
-        for (int i = 0; i < traces.length; i++) {
-            traces[i] = ScatterTrace.builder(x[i], y[i])
-            .mode(ScatterTrace.Mode.LINE)
-            .name(traceLabels[i])
-            .build();
-        }
-
-        Layout layout = Layout.builder()
-                .showLegend(true)
-                .xAxis(Axis.builder().title(xLabel).build())
-                .yAxis(Axis.builder().title(yLabel).build())
-                .build();
-
-        return new Figure(layout, traces);
-    }
-
-    public static Figure plotLogScale(double[][] x, double[][] y, String[] traceLabels, String xLabel, String yLabel) {
-        ScatterTrace[] traces = new ScatterTrace[x.length];
-        for (int i = 0; i < traces.length; i++) {
-            traces[i] = ScatterTrace.builder(x[i], y[i])
-            .mode(ScatterTrace.Mode.LINE)
-            .name(traceLabels[i])
-            .build();
-        }
-
-        Layout layout = Layout.builder()
-                .showLegend(true)
-                .xAxis(Axis.builder().type(Axis.Type.LOG).title(xLabel).build())
-                .yAxis(Axis.builder().type(Axis.Type.LOG).title(yLabel).build())
-                .build();
-
-        return new Figure(layout, traces);
-    }
 }
